requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
numpy>=1.20.0
scikit-learn>=0.24.0
tqdm>=4.62.0
tiktoken>=0.5.0
backoff>=2.2.0 